// src/app/models/UserDetailModel.ts
export class UserDetailModel {
  iduserDetail: number;
  dataNascimento: Date;
  dataIniciacao: Date;
  orixa: string;
  telefone: string;
  profissao: string;
  escolaridade: string;
  diaPagamento: Date;
  valorPagamento: number;
  endereco: string;
  numero: string;
  complemento: string;
  cidade: string;
  cep: string;
  estado: string;
  empregado: boolean;
  obs: string;
}